feat: prd-open-factory-session-selector-cleanup-007 - Add automated regression coverage for path validation and launch-target selection
